requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
alembic
python-dotenv
bcrypt
PyJWT
cryptography
pytz
marshmallow
flask_cors
flask_limiter
flask_swagger_ui